+++ src/main/java/org/realityforge/graphql/domgen/GraphQLSchemaService.java
@@ -1,0 +1,10 @@
+package org.realityforge.graphql.domgen;
+
+import graphql.servlet.GraphQLSchemaProvider;
+import javax.annotation.Nonnull;
+
+public interface GraphQLSchemaService
+{
+  @Nonnull
+  GraphQLSchemaProvider getProvider();
+}