+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 * Remove unused class `ValidationError`.
 * Remove unused class `DataFetchingError`.
 * Remove unused class `AbstractGraphQLError`.
+* Replace `AbstractGraphQLSchemaProvider` with `AbstractGraphQLSchemaService` that has a service method that
+  returns the `GraphQLSchemaProvider` instance.
+* Update `AbstractGraphQLSchemaService` to use graphqls to define schemas.
 
 ### [v1.3.0](https://github.com/realityforge/graphql-domgen-support/tree/v1.3.0) (2019-05-10)
 [Full Changelog](https://github.com/realityforge/graphql-domgen-support/compare/v1.2.0...v1.3.0)